fix: create a wrapper for `gh` which checks if we are in a jujutsu repositary and sets `GIT_DIR` accordingly
`gh` relies on the `.git` directory to determine repositary related
stuff. However, since we're using jujutsu for most repositories
`gh` is unable to locate the repository information. By checking
if we are in a jujutsu repositary and setting `GIT_DIR` to
`$JJ_ROOT/.jj/repo/store/git` we allow `gh` to find the necessary info.